requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-telegram-bot==21.4
replicate==0.32.0
sniffio==1.3.1
telegram==0.0.1
typing_extensions==4.12.2
